pyproject.toml
[project]
name = "matruspos"
version = "0.1.0"
description = "Plataforma de gerenciamento de disciplinas oferecidas via sistema Janus."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=2.18.2",
    "ipykernel>=6.29.5",
    "nodriver>=0.39",
    "pandas>=2.2.3",
    "playwright>=1.49.1",
    "python-dotenv>=1.0.1",
    "scrapy>=2.12.0",
    "selenium-driverless>=1.9.4",
    "websockets>=14.1",
]
